Bhai tum apna code sahi karo means apna project ka file structure sahi karo jitne bhi components banae hain unko components kae folder kae under rakho aur header missing hai aur bhi bhot sare chezzen hain first code ko clean likho
